Updated the property value converter to only include elements allowed by the data type
The property editor only allows selecting elements that has configured as allowed on the data type, but in situations where the element was selected at a time where the type was allowed, but then the data type configuration was changed to disallow the type, the element would still be returned as part of the property value.

With this commit, the property value converter will now check against the types allowed by the data type, and skip elements that are not allowed.